fix(plugin-workflow): fix client cycling import
