+++ packages/plugins/@nocobase/plugin-workflow/src/client/index.tsx
@@ -1,15 +1,3 @@
-export * from './Branch';
-export * from './FlowContext';
-export * from './constants';
-export * from './nodes';
-export { Trigger, useTrigger } from './triggers';
-export * from './variable';
-export * from './components';
-export * from './utils';
-export * from './hooks';
-export { default as useStyles } from './style';
-export * from './variable';
-
 import React from 'react';
 
 import { Plugin } from '@nocobase/client';
@@ -113,3 +101,16 @@ export default class PluginWorkflowClient extends Plugin {
     });
   }
 }
+
+export * from './Branch';
+export * from './FlowContext';
+export * from './constants';
+export * from './nodes';
+export { Trigger, useTrigger } from './triggers';
+export * from './variable';
+export * from './components';
+export * from './utils';
+export * from './hooks';
+export { default as useStyles } from './style';
+export * from './variable';
+export * from './ExecutionContextProvider';
